requirements.txt
git+https://github.com/Pycord-Development/pycord@slash
python-dotenv
tortoise-orm